// src/app/interfaces/riga.interface.ts
export interface IRiga {
    row_number: number;
    discriminante: string;
    informazioni_cronologiche: string;
    email: string;
    nome: string;
    cognome: string;
    luogo_nascita: string;
    data_nascita: string;
    numero_telefono: string;
    residenza: string;
    sei_occupata: string;
    che_lavoro_svolgi: string;
    quanti_anni_hai: string;
    diagnosi: string;
    dolore_mestruale: string;
    quanto_disturbante: string;
    da_quanto_tempo: string;
    riesci_a_gestire_dolore: string;
    come_gestisci_dolore: string;
    qualcuno_su_cui_contare: string;
    operata_per_endometriosi: string;
    qta_dolore_pelvico_cronico: string;
    qta_dolore_durante_rapporti: string;
    qta_dolori_intestinali: string;
    qta_dolori_schiena_lombari: string;
    qta_stanchezza_cronica: string;
    qta_cistiti_ricorrenti: string;
    qta_coliche: string;
    qta_cefalea: string;
    qta_gonfiore_addominale: string;
    qta_influire_sintomi_qualita_vita: string;
    assente_lavoro_scuola: string;
    interrompere_sport: string;
    rapporti_penetrativi: string;
    gia_paziente_endocare: string;
    se_si_chi_endocare: string;
    quali_specialisti_consultati: string;
    come_hai_conosciuto_endocare: string;
    perche_paziente_endocare: string;
    autorizzazione_trattamento_dati: string;
}